Switch to PKI type of encryption

+++ include/SimpleMon/msg.h
@@ -6,8 +6,8 @@
 struct StatusMessage {
 public:
     std::array <char, 32> hostname;
-    int uid = 0;
-    int free_mem = 0;
-    int free_disk = 0;
-    int used_cpu = 0;
+    short uid = 0;
+    short free_mem = 0;
+    short free_disk = 0;
+    short used_cpu = 0;
 };
